tambah role manager setengah

+++ app/Http/Controllers/DateClassController.php
@@ -3,8 +3,10 @@
 namespace App\Http\Controllers;
 
 use App\Models\DateClass;
+use App\Models\User;
 use Exception;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Yajra\DataTables\Facades\DataTables;
 
@@ -23,12 +25,15 @@ class DateClassController extends Controller
         $dataTable = DataTables::of($dates)
             ->addIndexColumn()
             ->addColumn('action', function ($data) {
-                $btn_action = '<div align="center">';
-                $btn_action .= '<a href="' . route('master.dateclass.edit', ['id' => $data->id]) . '" class="btn btn-sm btn-warning ml-2" title="Edit"><i class="fas fa-pencil-alt"></i></a>';
-                $btn_action .= '<button class="btn btn-sm btn-danger ml-2" onclick="destroyRecord(' . $data->id . ')" title="Delete"><i class="fas fa-trash"></i></button>';
+                if (User::find(Auth::user()->id)->hasRole('user')) {
+                    $btn_action = '<div align="center">';
+                    $btn_action .= '<a href="' . route('master.dateclass.edit', ['id' => $data->id]) . '" class="btn btn-sm btn-warning ml-2" title="Edit"><i class="fas fa-pencil-alt"></i></a>';
+                    $btn_action .= '<button class="btn btn-sm btn-danger ml-2" onclick="destroyRecord(' . $data->id . ')" title="Delete"><i class="fas fa-trash"></i></button>';
 
-                $btn_action .= '<div>';
-                return $btn_action;
+                    $btn_action .= '<div>';
+                    return $btn_action;
+                }
+                return null;
             })
 
             ->only(['name', 'date_code', 'action'])
